Clean up shop_name key on uninstall

+++ uninstall.php
@@ -1,3 +1,10 @@
 <?php
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) exit;
-delete_option( 'ds_toolkit_settings' );
+
+$opts = get_option( 'ds_toolkit_settings', array() );
+unset( $opts['shop_name'] );
+if ( empty( $opts ) ) {
+    delete_option( 'ds_toolkit_settings' );
+} else {
+    update_option( 'ds_toolkit_settings', $opts );
+}
